+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* `Connect-PartnerCenter -RefreshTokenScript`
+
 ## [0.0.1-alpha] - 2022-09-05
 * Release Test.